Properly apply radius increases of the CraftingStation and Ward to their EffectArea
CraftingStation and Wards have a child GameObject that has an EffectArea and a collider. The mob spawning system uses this collision to check whether spawning is allowed, and the collision radius is not automatically updated with the CraftingStation/Ward radii. So, without this, mobs can still spawn inside those expanded areas. By adjusting the EffectArea as well, we can ensure that any modification to the ranges there apply to mob spawning as well.