Make logo's "PLUS" use Norse font
The "PLUS" in the main menu logo in that round font and style and colors
didn't look quite good and kinda broke the aesthetics of the game.
Replaced with a white "PLUS" in Norse font (the rune-like font used in
the game), with a white drop shadow to make it look like it glows.