Add auto-split functionality on alt+click
Alt+click on a stack in a container will take as much of the stack as
you can carry without becoming encumbered.
If you already are encumbered, alt+click on a stack in your inventory
will move as *little* as possible to become unencumbered.

If you're in your inventory with no container open, alt+click will
instead *drop* as little as possible to become unencumbered.

Holding shift as well will toggle how the move happens.
The default is to "auto-move", which works essentially like ctrl+click
in the base game: it moves to partial stacks first, then to an empty slot
if still needed.

Holding alt+shift will split like the game's split feature and make the
user place the item(s) where they wish.